refactor(coordinator): decouple GateLocker via generic enum and relocate GateType to Generator
- GateLocker<TGate> generic over any struct Enum; drop ISettingProvider dependency
- Func<TGate, uint> replaces per-gate int limit; PerformanceSetting MaxParallel* → uint
- GateType moved from Coordinator.Domain to Generator.Domain
- IGateLocker<GateType> registered in Generator.Registration with per-gate uint lambda
- PrimaryEnlistment / SecondaryEnlistment → OwnerEnlistment / WaiterEnlistment
- Remove SlideGenerator.Collector module (orphan; superseded by Cloud)
- Normalization helper renamed to Naming in Utilities

+++ backend/tests/SlideGenerator.Settings.Tests/Unit/YamlSerializerTests.cs
@@ -70,11 +70,11 @@ public sealed class YamlSerializerTests
     public void Serialize_ThenDeserialize_CustomPerformanceSetting_PreservesValues()
     {
         var random = new Random();
-        var download = random.Next(1, 100);
-        var editImg = random.Next(1, 100);
-        var editPres = random.Next(1, 100);
-        var readWork = random.Next(1, 100);
-        var readPres = random.Next(1, 100);
+        var download = (uint)random.Next(1, 100);
+        var editImg = (uint)random.Next(1, 100);
+        var editPres = (uint)random.Next(1, 100);
+        var readWork = (uint)random.Next(1, 100);
+        var readPres = (uint)random.Next(1, 100);
 
         var original = new Setting
         {
